pyproject.toml
[project]
name = "python-di"
version = "0.0.1"
description = "A Python Dependency Injection Library"

dependencies=[
    "injector==0.22.0",
    "networkx==3.1",
    "numpy==1.24.3",
    "pydantic==2.11.3",
    "setuptools==68.0.0",
    "transformers==4.48.0",
    "python_util",
    "networkx==3.1",
    "numpy==1.24.3",
    "python-dotenv==1.1.0",
    "torch==2.7.0",
    "torch_geometric==2.6.1",
    "transformers==4.48.0",
    "typing_extensions==4.12.2",
    "pytest==8.3.4",
    "drools_py",
    "antlr4-python3-runtime==4.13.2"
]

keywords = ["commit-diff-context", "model-context-protocol"]

requires-python = ">=3.11"

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10"
]

[tool.uv.sources]
drools_py = { workspace = true }
python_util = { workspace = true }

[tool.setuptools]
packages = ["python_di"]
package-dir = {"" = "src"}

[tool.uv.build-backend]
source-include = ["*.yml"]

[tool.uv]
dependency-metadata = [
    { name = "python_di", source-include=["*.yml"] },
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
test-python-di = "python_di.test.run:main"

[tool.pyright]
venvPath = "."
venv = ".venv"